pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "dd-playground"
version = "0.1.0"
description = "Decision Diagram Playground for MQT Core"
authors = [{ name = "MQT" }]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mqt.core==3.0.2",
    "numpy==2.3.1",
    "matplotlib==3.10.3",
    "psutil==7.0.0",
    "ipywidgets==8.1.7",
    "jupyterlab==4.4.4",
    "pandas==2.2.2"
]

[tool.setuptools.package-dir]
"" = "src"

[tool.setuptools.packages.find]
where = ["src"]
